fix!: manage the `physical` interface in `ci-aio`
Treat the physical interface as network interface inside of the `ci-aio` environment. This
is to ensure the network interface files are created correctly with the appropriate permissions
otherwise we will lose connection after the DHCP lease expires. Also this approach enables us
to include the interface within the firewall configuration ensuring we maintain SSH access through
the firewall via this interface.